feat: isolate Money event constants from AppConstants dependencies
