+++ app/components/UI/Money/constants/moneyOnboardingEvents.ts
@@ -1,0 +1,14 @@
+export enum MONEY_ONBOARDING_STEP_ACTIONS {
+  // Generic actions
+  VIEWED = 'viewed',
+  SKIPPED = 'skipped',
+  EXITED = 'exited',
+  COMPLETED = 'completed',
+
+  // Transaction actions
+  DEPOSIT_INITIATED = 'deposit_initiated',
+
+  // Card actions
+  GET_CARD = 'get_card',
+  LINK_CARD = 'link_card',
+}
