+++ app/components/UI/Money/constants/moneyTooltipEvents.ts
@@ -1,0 +1,11 @@
+export enum MONEY_TOOLTIP_NAMES {
+  MONEY_BALANCE = 'money_balance',
+  MONTHLY_EARNINGS = 'monthly_earnings',
+  LIFETIME_EARNINGS = 'lifetime_earnings',
+  EARN_ON_YOUR_CRYPTO = 'earn_on_your_crypto',
+  APY = 'apy',
+}
+
+export enum MONEY_TOOLTIP_TYPES {
+  INFO = 'info',
+}
